requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic-settings
google-api-python-client
google-auth
google-cloud-secret-manager
gspread
